+++ src/Vitorm.Sqlite/DbContext_Extensions_UseSqlite.cs
@@ -1,6 +1,6 @@
 ﻿using Vitorm.Sql;
-using Vitorm.Sql.Transaction;
 using Vitorm.Sqlite;
+using Vitorm.Transaction;
 
 namespace Vitorm
 {
@@ -16,13 +16,13 @@ namespace Vitorm
                 dbConnectionProvider: config.ToDbConnectionProvider()
                 );
 
-            dbContext.createTransactionScope = createTransactionScope;
+            dbContext.createTransactionManager = createTransactionManager;
 
             if (config.commandTimeout.HasValue) dbContext.commandTimeout = config.commandTimeout.Value;
 
             return dbContext;
         }
 
-        static ITransactionScope createTransactionScope(SqlDbContext dbContext) => new Vitorm.Sqlite.SqlTransactionScope(dbContext);
+        static ITransactionManager createTransactionManager(SqlDbContext dbContext) => new Vitorm.Sqlite.SqlTransactionManager(dbContext);
     }
 }