+++ src/Vitorm/Sql/QueryExecutor/Async/ToListAsync.cs
@@ -38,7 +38,7 @@ namespace Vitorm.Sql.QueryExecutor
             var sql = sqlTranslateService.PrepareQuery(arg, combinedStream);
 
             // #3 Execute
-            using var reader = await dbContext.ExecuteReaderAsync(sql: sql, param: arg.sqlParam, useReadOnly: true);
+            using var reader = await dbContext.ExecuteReaderAsync(sql: sql, parameters: arg.sqlParam, useReadOnly: true);
             return (List<Result>)arg.dataReader.ReadData(reader);
         }