+++ src/Vitorm/Sql/QueryExecutor/Sync/ExecuteUpdate.cs
@@ -28,7 +28,7 @@ namespace Vitorm.Sql.QueryExecutor
             var sql = sqlTranslateService.PrepareExecuteUpdate(arg, streamToUpdate);
 
             // #3 Execute
-            return dbContext.Execute(sql: sql, param: arg.sqlParam);
+            return dbContext.Execute(sql: sql, parameters: arg.sqlParam);
         }
     }
 }