+++ src/Vitorm/Transaction/ITransaction.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Vitorm.Transaction
+{
+    public interface ITransaction : IDisposable
+    {
+        void Commit();
+        void Rollback();
+    }
+}