+++ doc/ReleaseNotes.md
@@ -1,6 +1,18 @@
 ﻿# Vitorm ReleaseNotes
 
 -----------------------
+# 2.4.0
+- [Vitorm] DbContext, new method GetEntityDescriptorFromCache
+- [Vitorm] move BeginTransaction to interface and refactor Transaction
+- [Vitorm] rename GetCurrentTransaction to GetDbTransaction
+- upgrade net to 8.0
+- [Vitorm] move SqlDataProvider to namespace Vitorm.Sql.DataProvider
+- [Vitorm] move IDbDataReader to namespace Vitorm.Sql.DataReader
+- [Vitorm] support Execute Procedure in SqlExecutor and SqlDbContext
+- [Vitorm] IDataReader extension methods: ReadValue ReadTuple ReadEntity
+
+
+-----------------------
 # 2.3.0
 
 - [Vitorm] support nested entity and rename Column to Property